Remove circular dependencies
`minimint` was depencency of `mint-client` and `ln-gateway`, which were
dependencies of `minimint`
