+++ ln-gateway/Cargo.toml
@@ -13,8 +13,9 @@ clap = { version = "3.1.18", features = ["derive"] }
 futures = "0.3.21"
 hex = "0.4.3"
 lightning-invoice = "0.14.0"
-minimint = { path = "../minimint" }
+minimint-shared = { path = "../minimint-shared" }
 minimint-api = { path = "../minimint-api" }
+minimint-ln = { path = "../modules/minimint-ln" }
 mint-client = { path = "../mint-client" }
 rand = "0.6"
 secp256k1 = "0.20"
