+++ minimint-shared/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "minimint-shared"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1.0.118", features = [ "derive" ] }
+serde_json = "1.0.61"
+bitcoin = "0.27.0"
+rand = "0.6.5"
+secp256k1-zkp = { git = "https://github.com/fedimint/rust-secp256k1-zkp", branch = "minimint", features = [ "global-context", "bitcoin_hashes" ] }
+
+# FIXME: is this appropriate for library crate?
+thiserror = "1.0.23"
+
+minimint-api = { path = "../minimint-api" }
+minimint-ln = { path = "../modules/minimint-ln" }
+minimint-mint = { path = "../modules/minimint-mint" }
+minimint-wallet = { path = "../modules/minimint-wallet" }
