+++ minimint-shared/src/outcome.rs
