+++ mint-client/Cargo.toml
@@ -18,8 +18,11 @@ hex = "0.4.3"
 lightning-invoice = "0.14.0"
 lightning = "0.0.106"
 miniscript = "6.0.0"
-minimint = { path = "../minimint" }
 minimint-api = { path = "../minimint-api" }
+minimint-shared = { path = "../minimint-shared" }
+minimint-wallet = { path = "../modules/minimint-wallet" }
+minimint-mint = { path = "../modules/minimint-mint" }
+minimint-ln = { path = "../modules/minimint-ln" }
 rand = "0.6.5"
 reqwest = { version = "0.11.0", features = [ "json" ], default-features = false }
 secp256k1-zkp = { git = "https://github.com/fedimint/rust-secp256k1-zkp", branch = "minimint", features = [ "serde", "bitcoin_hashes" ] }
