+++ mint-client/src/wallet/mod.rs
@@ -1,13 +1,13 @@
 use crate::BorrowedClientContext;
 use bitcoin::Address;
 use db::PegInKey;
-use minimint::config::FeeConsensus;
-use minimint::modules::wallet::config::WalletClientConfig;
-use minimint::modules::wallet::tweakable::Tweakable;
-use minimint::modules::wallet::txoproof::{PegInProof, PegInProofError, TxOutProof};
-use minimint::modules::wallet::PegOut;
 use minimint_api::db::batch::BatchTx;
 use minimint_api::Amount;
+use minimint_shared::FeeConsensus;
+use minimint_wallet::config::WalletClientConfig;
+use minimint_wallet::tweakable::Tweakable;
+use minimint_wallet::txoproof::{PegInProof, PegInProofError, TxOutProof};
+use minimint_wallet::PegOut;
 use miniscript::DescriptorTrait;
 use rand::{CryptoRng, RngCore};
 use secp256k1_zkp::schnorrsig::KeyPair;
@@ -138,17 +138,17 @@ mod tests {
     use async_trait::async_trait;
     use bitcoin::Address;
     use minimint::config::FeeConsensus;
-    use minimint::modules::ln::contracts::ContractId;
-    use minimint::modules::ln::ContractAccount;
-    use minimint::modules::wallet::bitcoind::test::{FakeBitcoindRpc, FakeBitcoindRpcController};
-    use minimint::modules::wallet::config::WalletClientConfig;
-    use minimint::modules::wallet::db::UTXOKey;
-    use minimint::modules::wallet::{SpendableUTXO, Wallet};
-    use minimint::outcome::{OutputOutcome, TransactionStatus};
-    use minimint::transaction::Transaction;
     use minimint_api::db::mem_impl::MemDatabase;
     use minimint_api::module::testing::FakeFed;
     use minimint_api::{Amount, OutPoint, TransactionId};
+    use minimint_ln::contracts::ContractId;
+    use minimint_ln::ContractAccount;
+    use minimint_shared::outcome::{OutputOutcome, TransactionStatus};
+    use minimint_shared::transaction::Transaction;
+    use minimint_wallet::bitcoind::test::{FakeBitcoindRpc, FakeBitcoindRpcController};
+    use minimint_wallet::config::WalletClientConfig;
+    use minimint_wallet::db::UTXOKey;
+    use minimint_wallet::{SpendableUTXO, Wallet};
     use std::str::FromStr;
     use std::sync::Arc;
     use std::time::Duration;
